refactor: 엔티티 웹툰 좋아요 구조 변경
